feat: refactor DateRangePicker and DateTimePicker for improved date handling and UI interactions
